+++ app/commands/homebrew_context/daggerheart/add_book_items_command.rb
@@ -2,7 +2,7 @@
 
 module HomebrewContext
   module Daggerheart
-    class AddBookCommunitiesCommand < BaseCommand
+    class AddBookItemsCommand < BaseCommand
       use_contract do
         config.messages.namespace = :homebrew_book
 
@@ -16,17 +16,18 @@ module HomebrewContext
       private
 
       def do_prepare(input)
-        existing_community_ids = input[:book].items.where(itemable_type: 'Daggerheart::Homebrew::Community').pluck(:itemable_id)
-        community_ids =
-          Homebrew::Community
-            .where(type: 'Daggerheart::Homebrew::Community', user_id: input[:user].id, name: input[:names])
-            .where.not(id: existing_community_ids)
+        existing_items_ids = input[:book].items.where(itemable_type: 'Daggerheart::Item').pluck(:itemable_id)
+        items_ids =
+          ::Item
+            .where(type: 'Daggerheart::Item', user_id: input[:user].id)
+            .where("name ->> 'en' = :name", name: input[:names])
+            .where.not(id: existing_items_ids)
             .ids
-        input[:attributes] = community_ids.map do |community_id|
+        input[:attributes] = items_ids.map do |item_id|
           {
             homebrew_book_id: input[:book].id,
-            itemable_type: 'Daggerheart::Homebrew::Community',
-            itemable_id: community_id
+            itemable_type: 'Daggerheart::Item',
+            itemable_id: item_id
           }
         end
       end