+++ app/commands/homebrew_context/daggerheart/add_book_subclasses_command.rb
@@ -2,7 +2,7 @@
 
 module HomebrewContext
   module Daggerheart
-    class AddBookCommunitiesCommand < BaseCommand
+    class AddBookSubclassesCommand < BaseCommand
       use_contract do
         config.messages.namespace = :homebrew_book
 
@@ -16,17 +16,17 @@ module HomebrewContext
       private
 
       def do_prepare(input)
-        existing_community_ids = input[:book].items.where(itemable_type: 'Daggerheart::Homebrew::Community').pluck(:itemable_id)
-        community_ids =
-          Homebrew::Community
-            .where(type: 'Daggerheart::Homebrew::Community', user_id: input[:user].id, name: input[:names])
-            .where.not(id: existing_community_ids)
+        existing_subclasses_ids = input[:book].items.where(itemable_type: 'Daggerheart::Homebrew::Subclass').pluck(:itemable_id)
+        subclasses_ids =
+          Homebrew::Subclass
+            .where(type: 'Daggerheart::Homebrew::Subclass', user_id: input[:user].id, name: input[:names])
+            .where.not(id: existing_subclasses_ids)
             .ids
-        input[:attributes] = community_ids.map do |community_id|
+        input[:attributes] = subclasses_ids.map do |subclass_id|
           {
             homebrew_book_id: input[:book].id,
-            itemable_type: 'Daggerheart::Homebrew::Community',
-            itemable_id: community_id
+            itemable_type: 'Daggerheart::Homebrew::Subclass',
+            itemable_id: subclass_id
           }
         end
       end