+++ app/javascript/applications/CharKeeperApp/pages/Content/Homebrew/Modules.jsx
@@ -53,6 +53,18 @@ export const HomebrewModules = (props) => {
                         <p class="flex-1">{community.name}</p>
                       }
                     </For>
+                    <p class="mt-8 mb-2">{t(`pages.homebrewPage.${props.provider}.includedSubclasses`)}</p>
+                    <For each={props.homebrews.subclasses.filter((item) => book.items.subclasses.includes(item.id))}>
+                      {(subclass) =>
+                        <p class="flex-1">{subclass.name}</p>
+                      }
+                    </For>
+                    <p class="mt-8 mb-2">{t(`pages.homebrewPage.${props.provider}.includedItems`)}</p>
+                    <For each={props.homebrews.items.filter((item) => book.items.items.includes(item.id))}>
+                      {(item) =>
+                        <p class="flex-1">{item.name.en}</p>
+                      }
+                    </For>
                   </Match>
                 </Switch>
               </Show>