+++ app/services/bot_context/commands/book.rb
@@ -7,9 +7,12 @@ module BotContext
         add_book_command: 'commands.homebrew_context.add_book',
         remove_book_command: 'commands.homebrew_context.remove_book',
         add_book_races_command: 'commands.homebrew_context.daggerheart.add_book_races',
-        add_book_communities_command: 'commands.homebrew_context.daggerheart.add_book_communities'
+        add_book_communities_command: 'commands.homebrew_context.daggerheart.add_book_communities',
+        add_book_subclasses_command: 'commands.homebrew_context.daggerheart.add_book_subclasses',
+        add_book_items_command: 'commands.homebrew_context.daggerheart.add_book_items'
       ]
 
+      # rubocop: disable Metrics/CyclomaticComplexity
       def call(source:, arguments:, data:)
         case arguments.shift
         when 'create' then create_book(*arguments, data, source)
@@ -19,9 +22,12 @@ module BotContext
         when 'set' then set_book(*arguments, data, source)
         when 'addRace' then add_race(*arguments, data, source)
         when 'addCommunity' then add_community(*arguments, data, source)
+        when 'addSubclass' then add_subclass(*arguments, data, source)
+        when 'addItem' then add_item(*arguments, data, source)
         when 'export' then export_book(data, source)
         end
       end
+      # rubocop: enable Metrics/CyclomaticComplexity
 
       private
 
@@ -90,6 +96,26 @@ module BotContext
         add_book_communities_command.call(user: data[:user], book: ::Homebrew::Book.find(object.info['id']), names: arguments)
         {
           type: 'add_community',
+          result: :ok,
+          errors: nil
+        }
+      end
+
+      def add_subclass(*arguments, data, source)
+        object = ActiveBotObject.find_by!(user: data[:user], source: source, object: 'book')
+        add_book_subclasses_command.call(user: data[:user], book: ::Homebrew::Book.find(object.info['id']), names: arguments)
+        {
+          type: 'add_subclass',
+          result: :ok,
+          errors: nil
+        }
+      end
+
+      def add_item(*arguments, data, source)
+        object = ActiveBotObject.find_by!(user: data[:user], source: source, object: 'book')
+        add_book_items_command.call(user: data[:user], book: ::Homebrew::Book.find(object.info['id']), names: arguments)
+        {
+          type: 'add_item',
           result: :ok,
           errors: nil
         }