+++ config/initializers/container.rb
@@ -92,6 +92,10 @@ module Charkeeper
     register('commands.homebrew_context.daggerheart.add_book_communities') {
       HomebrewContext::Daggerheart::AddBookCommunitiesCommand.new
     }
+    register('commands.homebrew_context.daggerheart.add_book_subclasses') {
+      HomebrewContext::Daggerheart::AddBookSubclassesCommand.new
+    }
+    register('commands.homebrew_context.daggerheart.add_book_items') { HomebrewContext::Daggerheart::AddBookItemsCommand.new }
 
     register('commands.campaigns_context.add_campaign') { CampaignsContext::AddCampaignCommand.new }
     register('commands.campaigns_context.join_campaign') { CampaignsContext::JoinCampaignCommand.new }